requirements.txt
#--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.8.1/constraints-3.11.txt"
boto3==1.28.0
requests==2.31.0
pandas==2.1.2
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-sqlalchemy